change DataType for spotPrice to better reflect AirBnB

+++ backend/db/migrations/20221023175930-create-spot.js
@@ -49,7 +49,7 @@ module.exports = {
         allowNull: false
       },
       price: {
-        type: Sequelize.DECIMAL(7, 2),
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       createdAt: {
